refactor: 패키지 org.example를 burgermap으로 변경
- 패키지 이름을 프로젝트 정체성을 드러내는 이름으로 변경
