+++ api/src/main/java/burgermap/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package burgermap;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
