+++ core/src/main/java/burgermap/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package burgermap;
 
 public class Main {
     public static void main(String[] args) {
